refactor: 입출력 패키지 분리
